Pause Menu Refactor

+++ source/Tween.cpp
@@ -9,25 +9,25 @@ float Tween::linInterp(float start, float end, float percentage) {
 	return start * (1.0f - percentage) + end * percentage;
 }
 
-float Tween::linear(float start, float end, int currFrame, int maxFrame) {
+float Tween::linear(float start, float end, size_t currFrame, size_t maxFrame) {
 	return linInterp(start, end, static_cast<float>(currFrame) / static_cast<float>(maxFrame));
 }
 
-float Tween::easeIn(float start, float end, int currFrame, int maxFrame) {
+float Tween::easeIn(float start, float end, size_t currFrame, size_t maxFrame) {
 	float t = static_cast<float>(currFrame) / static_cast<float>(maxFrame);
 	return linInterp(start, end, t * t * t * t);
 }
 
-float Tween::easeOut(float start, float end, int currFrame, int maxFrame) {
+float Tween::easeOut(float start, float end, size_t currFrame, size_t maxFrame) {
 	float t = static_cast<float>(currFrame) / static_cast<float>(maxFrame);
 	t--;
 	return linInterp(start, end, -t * t * t * t + 1);
 }
 
-float Tween::easeInOut(float start, float end, int currFrame, int maxFrame) {
+float Tween::easeInOut(float start, float end, size_t currFrame, size_t maxFrame) {
 	const float t = static_cast<float>(currFrame) / static_cast<float>(maxFrame);
 	const float halfPos = ((end - start) / 2) + start;
-	const int halfFrame = maxFrame / 2;
+	const size_t halfFrame = maxFrame / 2;
 	if (2 * t < 1) {
 		return easeIn(start, halfPos, currFrame, halfFrame);
 	}
